requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
marshmallow==3.20.2
python-dotenv==1.0.1
boto3==1.34.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
requests==2.31.0
pytest
pytest-flask
pytest-mock
python-json-logger